feat(appwrite): add appwrite.json template
Developers can use this template for their appwrite.json
to deploy the project.

The appwrite.json file should be ignored from git because
it may contain sensitive values like function variables.

The appwrite.template.json should be ignored by biome so
that it isn't reformatted, making it out of sync with
the appwrite.json file.